Michael

+++ App A/src/MMBOS/Accounts.java
@@ -1,11 +1,11 @@
 package MMBOS;
 
 public class Accounts {
-    public int accountNumber;
+    public long accountNumber;
     public String personalID;
     public double cashInAccount;
 
-    public Accounts(int accountNumber, String personalID, double cashInAccount) {
+    public Accounts(long accountNumber, String personalID, double cashInAccount) {
         this.accountNumber = accountNumber;
         this.personalID = personalID;
         this.cashInAccount = cashInAccount;
